Fill segment data with segment number
Track the current segment number for each connection in the server, and
fill the "dummy" video data packets with that value (truncated to 8
bits).

Track the current segment number in the client and verify packet
contents match that segment number.

This makes it easier to detect large out-of-order delivery errors which
span video segments and to see program behaviour in wireshark without a
disector for QUIC.